fix(dashboard): keep the embed marker sticky across navigation
Review hardening, second round: once the Desktop handoff parameter has
been seen, the dashboard stays embedded for the session — internal
navigation routes and daemon-origin redirects rebuild the URL without
desktop_embed, which would otherwise re-expose the app-racing Update
Guard button inside the Desktop window. A recovery-reinstall prompt now
points embedded readers at the app's updater before suggesting a
terminal repair, the quarantine sweep rechecks each sidecar for symlink
swaps before unlinking, and the pure copy helpers moved to their own
module to stay under the size ratchet.
