+++ dashboard/src/desktop-embed.ts
@@ -4,10 +4,18 @@
  * Updates"), so the embedded dashboard must not offer a second, competing
  * update action against the same runtime.
  */
+let embeddedLatch = false;
 let memoForHref: string | null = null;
 let memoResult = false;
 
 export function dashboardEmbedsInDesktop(): boolean {
+  // Once the Desktop handoff is seen it stays true for the session: internal
+  // navigation replaces the URL with routes that no longer carry the
+  // desktop_embed parameter, and a daemon-origin redirect rebuilds the URL
+  // the same way — neither turns an embedded dashboard standalone.
+  if (embeddedLatch) {
+    return true;
+  }
   let href: string;
   try {
     href = window.location.href;
@@ -28,5 +36,6 @@ export function dashboardEmbedsInDesktop(): boolean {
   }
   memoForHref = href;
   memoResult = params.get("desktop_embed") === "1";
+  embeddedLatch = memoResult;
   return memoResult;
 }
