+++ dashboard/src/guard-update-panel.tsx
@@ -14,6 +14,7 @@ import {
 } from "./guard-api";
 import { AlphaChannelDialog } from "./alpha-update-channel-dialog";
 import { dashboardEmbedsInDesktop } from "./desktop-embed";
+import { updateStatusLabel, updateHelpCopy } from "./guard-update-copy";
 import { buildApprovalProofCredentials } from "./approval-proof-inline";
 import type {
   GuardApprovalGatePublicConfig,
@@ -39,83 +40,6 @@ export type GuardUpdatePanelProps = {
   onSetUpdateChannel?: (channel: "stable" | "alpha", proof?: GuardUpdateChannelProof) => void | Promise<void>;
   compact?: boolean;
 };
-
-function updateStatusLabel(status: GuardUpdateStatus | null | undefined): string {
-  if (!status) {
-    return "Checking version…";
-  }
-  if (status.update_available && status.latest_version) {
-    return `Version ${status.latest_version} is ready`;
-  }
-  return `Version ${status.current_version}`;
-}
-
-function shouldPromptRecoveryReinstall(status: GuardUpdateStatus | null | undefined): boolean {
-  return (
-    status?.recovery_reinstall_available === true &&
-    status?.auto_updatable !== true &&
-    status?.version_check?.update_available === true
-  );
-}
-
-function recoveryReinstallHelpCopy(status: GuardUpdateStatus | null | undefined): string | null {
-  if (!shouldPromptRecoveryReinstall(status)) {
-    return null;
-  }
-  const blockedReason = status?.blocked_reason ?? "";
-  if (blockedReason.includes("local wheel whose source file is no longer available")) {
-    return "This install came from a local wheel whose source file is no longer available, so automatic updates are off. Reinstall from PyPI to switch it back to a normal package; Guard restarts briefly and saved approvals stay.";
-  }
-  if (blockedReason.includes("local wheel")) {
-    return "This install came from a local wheel, so automatic updates are off. Reinstall from PyPI to switch it back to a normal package; Guard restarts briefly and saved approvals stay.";
-  }
-  return "This install came from a local folder, so automatic updates are off. Reinstall from PyPI to switch it back to a normal package; Guard restarts briefly and saved approvals stay.";
-}
-
-function updateHelpCopy(
-  status: GuardUpdateStatus | null | undefined,
-  phase: GuardUpdatePhase,
-  errorMessage?: string | null,
-  embeddedInDesktop = false,
-): string | null {
-  if (phase === "updating") {
-    return "Guard is installing the update. The dashboard will pause briefly and reopen when ready.";
-  }
-  if (phase === "reconnecting") {
-    return "Reconnecting to Guard after the update…";
-  }
-  if (phase === "error") {
-    if (embeddedInDesktop) {
-      return (
-        errorMessage?.trim() ||
-        "The update did not finish. Use Check for Updates in the HOL Guard menu bar and watch its progress there."
-      );
-    }
-    return errorMessage?.trim() || "The update did not finish. The installed version stays in place. Try again, or run hol-guard update from your terminal.";
-  }
-  if (status?.update_suppressed) {
-    if (status.retry_command) {
-      return `Automatic update already ran but this install is still behind. Run ${status.retry_command} in your terminal.`;
-    }
-    if (status.update_attempt_message) {
-      return status.update_attempt_message;
-    }
-    return "Automatic update already ran but this install is still behind the latest release.";
-  }
-  if (status?.update_available) {
-    if (embeddedInDesktop) {
-      return "Updates run through the HOL Guard app. Use Check for Updates in the HOL Guard menu-bar icon, and the app installs this version with its own progress screen.";
-    }
-    return "This restarts Guard for a moment. Open approvals will stay saved.";
-  }
-  if (status && !status.auto_updatable && status.recovery_reinstall_available) {
-    return recoveryReinstallHelpCopy(status);
-  }
-  if (status && !status.auto_updatable && status.blocked_reason) {
-    return status.blocked_reason;
-  }
-  return null;
-}
 
 export function GuardUpdatePanel(props: GuardUpdatePanelProps) {
   const version = props.guardVersion ?? props.updateStatus?.current_version ?? null;
